can route to an album by id in the url
