+++ src/pages/AlbumPhotos.jsx
@@ -1,13 +1,36 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import { Link, useLocation} from 'react-router-dom';
 
 export default function AlbumPhotos(props) {
     let location = useLocation();
-    let id = location.state.albumId;
-    let photos = location.state.albumPhotos;
-
+    let searchParams = new URLSearchParams(location.search);
+    let id = Number(searchParams.get("albumId"));
+    
+    console.log(typeof(id))
     const [albumsImage, setalbumsImage] = useState("https://via.placeholder.com/600/92c952");
 
+    const [photos, setPhotos] = useState([]);
+
+    useEffect(() => {
+      const getPhotos = async () => {
+        try {
+          let photosResponse = await fetch("https://jsonplaceholder.typicode.com/photos");
+    
+          if (!photosResponse.ok) {
+            throw new Error("Error, please try again");
+          }
+    
+          let photosData = await photosResponse.json();
+          setPhotos(photosData);
+        } catch (error) {
+          console.error("There has been a problem with your fetch operation:", error);
+        }
+      };
+    
+      getPhotos();
+      
+    }, []);
+    console.log("ID: ", id)
 
   return (
     <div>
@@ -19,8 +42,10 @@ export default function AlbumPhotos(props) {
             <div style={{display: "flex", flexDirection: "row", overflowX: "scroll",}}>
             {
               photos.map((photo, index) => {
-                if(id === photo.albumId)
+                if(id === photo.albumId){
+                  console.log("HEYYYYY")
                   return <img src={photo.thumbnailUrl} key={index} onClick={() => setalbumsImage(photo.url)}></img>
+                }
               })
             }
             </div>
